+++ packages/sdk/src/adaptors/rss3.ts
@@ -2,26 +2,17 @@ import { AdaptorDeets, Dictionary } from '../types';
 import { fetcher } from '../utils';
 
 interface Rss3QueryResult {
-  profile: Dictionary<string>;
-  '@backlinks': Array<Dictionary<string>>;
-  accounts: Array<Dictionary<string>>;
-  links: Array<Dictionary<string>>;
+  total: number,
+  result: Array<string>
 }
 
 export default async function getRss3Data(address: string) {
   const jsonData = (await fetcher(
     'GET',
-    `https://hub.pass3.me/${address}`
+    `https://pregod.rss3.dev/v1/profiles/${address}`
   )) as Rss3QueryResult;
 
-  return {
-    profile: Boolean(jsonData['profile']) === true ? jsonData['profile'] : {},
-    backlinks:
-      Boolean(jsonData['@backlinks']) === true ? jsonData['@backlinks'] : [],
-    accounts:
-      Boolean(jsonData['accounts']) === true ? jsonData['accounts'] : [],
-    links: Boolean(jsonData['links']) === true ? jsonData['links'] : [],
-  };
+  return jsonData?.result?.length > 0 ? jsonData?.result : false
 }
 
 export const Rss3AdaptorDeets: AdaptorDeets = {
@@ -29,6 +20,6 @@ export const Rss3AdaptorDeets: AdaptorDeets = {
   name: 'Rss3',
   projectThumbnail:
     'ipfs://bafybeidgse3grhvqzcp23ndji4hxtnd4x3bjt7r5jel3zwdqcghyhc2oim/rss3.webp',
-  projectUrl: 'https://roci.io/',
+  projectUrl: 'https://rss3.io/',
   requiredConfigKeys: [],
 };