+++ packages/sdk/src/adaptors/tokenblacklists.ts
@@ -8,26 +8,30 @@ interface TokenBlacklistResult {
 }
 
 export default async function getTokenBlacklistData(address: string) {
-  const { data } = (await gqlFetcher(
-    'https://api.studio.thegraph.com/query/1649/token-blacklists/v1.6',
-    `{
-        blacklist(id: "${address.toLowerCase()}") {
-          alusd
-          busd
-          euroc
-          eurt
-          husd
-          tusd
-          usdc
-          usdp
-          usdt
-          lastUpdated
-        }
-    }`
-  )) as TokenBlacklistResult;
+  try {
+    const { data } = (await gqlFetcher(
+      'https://api.studio.thegraph.com/query/1649/token-blacklists/v1.6',
+      `{
+          blacklist(id: "${address.toLowerCase()}") {
+            alusd
+            busd
+            euroc
+            eurt
+            husd
+            tusd
+            usdc
+            usdp
+            usdt
+            lastUpdated
+          }
+      }`
+    )) as TokenBlacklistResult;
 
-  if (data.blacklist === null) return false;
-  else return data.blacklist;
+    if (data?.blacklist === null) return false;
+    else return data?.blacklist;
+  } catch (error) {
+    return false;
+  }
 }
 
 export const TokenblacklistsAdaptorDeets: AdaptorDeets = {